Removes support cases functions to own file
This PR moves functions for AWS support cases-related code to their own file. This is the first of several to reduce the length and complexity of the accounts_controller.go file.

REF: OSD-1937 "aws-account-operator breakout functions from account_controller.go"
